Renames a module
