+++ js/main.js
@@ -1,4 +1,4 @@
-import { getData, showAlert } from './api.js';
+import { getData, showAlert } from './creates-api.js';
 import { renderGallery } from './renders-gallery.js';
 import {userFotoFormSubmit} from './upload-foto.js';
